+++ languagetool-language-modules/km/src/main/java/org/languagetool/language/Khmer.java
@@ -61,7 +61,13 @@ public class Khmer extends Language {
   public String[] getCountries() {
     return new String[]{"KH"};
   }
-  
+
+  @Override
+  public String getCommonWordsPath() {
+    // TODO: provide common words file
+    return null;
+  }
+
   @Override
   public Tagger getTagger() {
     if (tagger == null) {